+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-04 10:37:10
+
+### Changed Files
+- `src/features/policies/PolicyForm.tsx`
+- `src/features/policies/components/PolicyFormPolicySection.tsx`
+- `src/features/policies/components/SubmitDateConfirmDialog.tsx`
+- `src/features/policies/hooks/usePolicyForm.ts`
+
+### Statistics
+```
+ 4 files changed, 278 insertions(+), 23 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-04 10:20:21
 
 ### Changed Files